+++ DataAccess/Models/Metal.cs
@@ -5,7 +5,7 @@ namespace DataAccess.Models;
 
 public partial class Metal
 {
-    public short Gkey { get; set; }
+    public int Gkey { get; set; }
 
     public string? MetalName { get; set; }
 }
